Split product types file
