+++ saleor/graphql/product/types/__init__.py
@@ -1,13 +1,11 @@
+from .categories import Category, CategoryCountableConnection
+from .collections import Collection, CollectionCountableConnection
 from .digital_contents import (
     DigitalContent,
     DigitalContentCountableConnection,
     DigitalContentUrl,
 )
 from .products import (
-    Category,
-    CategoryCountableConnection,
-    Collection,
-    CollectionCountableConnection,
     Product,
     ProductCountableConnection,
     ProductMedia,
