+++ saleor/graphql/product/types/products.py
@@ -103,12 +103,7 @@ from ...tax.dataloaders import (
 )
 from ...tax.types import TaxClass
 from ...translations.fields import TranslationField
-from ...translations.types import (
-    CategoryTranslation,
-    CollectionTranslation,
-    ProductTranslation,
-    ProductVariantTranslation,
-)
+from ...translations.types import ProductTranslation, ProductVariantTranslation
 from ...utils import get_user_or_app_from_context
 from ...utils.filters import reporting_period_to_date
 from ...warehouse.dataloaders import (
@@ -119,9 +114,7 @@ from ...warehouse.dataloaders import (
 from ...warehouse.types import Stock
 from ..dataloaders import (
     CategoryByIdLoader,
-    CategoryChildrenByCategoryIdLoader,
     CollectionChannelListingByCollectionIdAndChannelSlugLoader,
-    CollectionChannelListingByCollectionIdLoader,
     CollectionsByProductIdLoader,
     ImagesByProductIdLoader,
     ImagesByProductVariantIdLoader,
@@ -136,8 +129,6 @@ from ..dataloaders import (
     ProductVariantsByProductIdLoader,
     SelectedAttributesByProductIdLoader,
     SelectedAttributesByProductVariantIdLoader,
-    ThumbnailByCategoryIdSizeAndFormatLoader,
-    ThumbnailByCollectionIdSizeAndFormatLoader,
     ThumbnailByProductMediaIdSizeAndFormatLoader,
     VariantAttributesByProductTypeIdLoader,
     VariantChannelListingByVariantIdAndChannelSlugLoader,
@@ -145,14 +136,8 @@ from ..dataloaders import (
     VariantsChannelListingByProductIdAndChannelSlugLoader,
 )
 from ..enums import ProductMediaType, ProductTypeKindEnum, VariantAttributeScope
-from ..filters import ProductFilterInput
 from ..resolvers import resolve_product_variants, resolve_products
-from ..sorters import ProductOrder
-from .channels import (
-    CollectionChannelListing,
-    ProductChannelListing,
-    ProductVariantChannelListing,
-)
+from .channels import ProductChannelListing, ProductVariantChannelListing
 from .digital_contents import DigitalContent
 
 destination_address_argument = graphene.Argument(
@@ -819,7 +804,7 @@ class Product(ChannelContextTypeWithMetadata, ModelObjectType):
     description = JSONString(description="Description of the product." + RICH_CONTENT)
     product_type = graphene.Field(lambda: ProductType, required=True)
     slug = graphene.String(required=True)
-    category = graphene.Field(lambda: Category)
+    category = graphene.Field("saleor.graphql.product.types.categories.Category")
     created = graphene.DateTime(required=True)
     updated_at = graphene.DateTime(required=True)
     charge_taxes = graphene.Boolean(
@@ -903,7 +888,7 @@ class Product(ChannelContextTypeWithMetadata, ModelObjectType):
         deprecation_reason=f"{DEPRECATED_IN_3X_FIELD} Use the `media` field instead.",
     )
     collections = NonNullList(
-        lambda: Collection,
+        "saleor.graphql.product.types.collections.Collection",
         description=(
             "List of collections for the product. Requires the following permissions "
             "to include the unpublished items: "
@@ -1670,266 +1655,6 @@ class ProductType(ModelObjectType):
 class ProductTypeCountableConnection(CountableConnection):
     class Meta:
         node = ProductType
-
-
-@federated_entity("id channel")
-class Collection(ChannelContextTypeWithMetadata, ModelObjectType):
-    id = graphene.GlobalID(required=True)
-    seo_title = graphene.String()
-    seo_description = graphene.String()
-    name = graphene.String(required=True)
-    description = JSONString(
-        description="Description of the collection." + RICH_CONTENT
-    )
-    slug = graphene.String(required=True)
-    channel = graphene.String(
-        description=(
-            "Channel given to retrieve this collection. Also used by federation "
-            "gateway to resolve this object in a federated query."
-        ),
-    )
-    description_json = JSONString(
-        description="Description of the collection." + RICH_CONTENT,
-        deprecation_reason=(
-            f"{DEPRECATED_IN_3X_FIELD} Use the `description` field instead."
-        ),
-    )
-    products = FilterConnectionField(
-        ProductCountableConnection,
-        filter=ProductFilterInput(description="Filtering options for products."),
-        sort_by=ProductOrder(description="Sort products."),
-        description="List of products in this collection.",
-    )
-    background_image = ThumbnailField()
-    translation = TranslationField(
-        CollectionTranslation,
-        type_name="collection",
-        resolver=ChannelContextType.resolve_translation,
-    )
-    channel_listings = PermissionsField(
-        NonNullList(CollectionChannelListing),
-        description="List of channels in which the collection is available.",
-        permissions=[
-            ProductPermissions.MANAGE_PRODUCTS,
-        ],
-    )
-
-    class Meta:
-        default_resolver = ChannelContextType.resolver_with_context
-        description = "Represents a collection of products."
-        interfaces = [relay.Node, ObjectWithMetadata]
-        model = models.Collection
-
-    @staticmethod
-    def resolve_channel(root: ChannelContext[models.Product], _info):
-        return root.channel_slug
-
-    @staticmethod
-    def resolve_background_image(
-        root: ChannelContext[models.Collection], info, size=None, format=None
-    ):
-        node = root.node
-        if not node.background_image:
-            return
-
-        alt = node.background_image_alt
-        if not size:
-            return Image(url=node.background_image.url, alt=alt)
-
-        format = format.lower() if format else None
-        size = get_thumbnail_size(size)
-
-        def _resolve_background_image(thumbnail):
-            url = get_image_or_proxy_url(thumbnail, node.id, "Collection", size, format)
-            return Image(url=url, alt=alt)
-
-        return (
-            ThumbnailByCollectionIdSizeAndFormatLoader(info.context)
-            .load((node.id, size, format))
-            .then(_resolve_background_image)
-        )
-
-    @staticmethod
-    def resolve_products(root: ChannelContext[models.Collection], info, **kwargs):
-        requestor = get_user_or_app_from_context(info.context)
-        qs = root.node.products.visible_to_user(  # type: ignore
-            requestor, root.channel_slug
-        )
-        qs = ChannelQsContext(qs=qs, channel_slug=root.channel_slug)
-
-        kwargs["channel"] = root.channel_slug
-        qs = filter_connection_queryset(qs, kwargs)
-        return create_connection_slice(qs, info, kwargs, ProductCountableConnection)
-
-    @staticmethod
-    def resolve_channel_listings(root: ChannelContext[models.Collection], info):
-        return CollectionChannelListingByCollectionIdLoader(info.context).load(
-            root.node.id
-        )
-
-    @staticmethod
-    def resolve_description_json(root: ChannelContext[models.Collection], _info):
-        description = root.node.description
-        return description if description is not None else {}
-
-    @staticmethod
-    def __resolve_references(roots: List["Collection"], info):
-        from ..resolvers import resolve_collections
-
-        channels = defaultdict(set)
-        roots_ids = []
-        for root in roots:
-            _, root_id = from_global_id_or_error(root.id, Collection, raise_error=True)
-            roots_ids.append(f"{root.channel}_{root_id}")
-            channels[root.channel].add(root_id)
-
-        collections = {}
-        for channel, ids in channels.items():
-            queryset = resolve_collections(info, channel).qs.filter(id__in=ids)
-
-            for collection in queryset:
-                collections[f"{channel}_{collection.id}"] = ChannelContext(
-                    channel_slug=channel, node=collection
-                )
-
-        return [collections.get(root_id) for root_id in roots_ids]
-
-
-class CollectionCountableConnection(CountableConnection):
-    class Meta:
-        node = Collection
-
-
-@federated_entity("id")
-class Category(ModelObjectType):
-    id = graphene.GlobalID(required=True)
-    seo_title = graphene.String()
-    seo_description = graphene.String()
-    name = graphene.String(required=True)
-    description = JSONString(description="Description of the category." + RICH_CONTENT)
-    slug = graphene.String(required=True)
-    parent = graphene.Field(lambda: Category)
-    level = graphene.Int(required=True)
-    description_json = JSONString(
-        description="Description of the category." + RICH_CONTENT,
-        deprecation_reason=(
-            f"{DEPRECATED_IN_3X_FIELD} Use the `description` field instead."
-        ),
-    )
-    ancestors = ConnectionField(
-        lambda: CategoryCountableConnection,
-        description="List of ancestors of the category.",
-    )
-    products = ConnectionField(
-        ProductCountableConnection,
-        channel=graphene.String(
-            description="Slug of a channel for which the data should be returned."
-        ),
-        description=(
-            "List of products in the category. Requires the following permissions to "
-            "include the unpublished items: "
-            f"{', '.join([p.name for p in ALL_PRODUCTS_PERMISSIONS])}."
-        ),
-    )
-    children = ConnectionField(
-        lambda: CategoryCountableConnection,
-        description="List of children of the category.",
-    )
-    background_image = ThumbnailField()
-    translation = TranslationField(CategoryTranslation, type_name="category")
-
-    class Meta:
-        description = (
-            "Represents a single category of products. Categories allow to organize "
-            "products in a tree-hierarchies which can be used for navigation in the "
-            "storefront."
-        )
-        interfaces = [relay.Node, ObjectWithMetadata]
-        model = models.Category
-
-    @staticmethod
-    def resolve_ancestors(root: models.Category, info, **kwargs):
-        return create_connection_slice(
-            root.get_ancestors(), info, kwargs, CategoryCountableConnection
-        )
-
-    @staticmethod
-    def resolve_description_json(root: models.Category, _info):
-        description = root.description
-        return description if description is not None else {}
-
-    @staticmethod
-    def resolve_background_image(root: models.Category, info, size=None, format=None):
-        if not root.background_image:
-            return
-
-        alt = root.background_image_alt
-        if not size:
-            return Image(url=root.background_image.url, alt=alt)
-
-        format = format.lower() if format else None
-        size = get_thumbnail_size(size)
-
-        def _resolve_background_image(thumbnail):
-            url = get_image_or_proxy_url(thumbnail, root.id, "Category", size, format)
-            return Image(url=url, alt=alt)
-
-        return (
-            ThumbnailByCategoryIdSizeAndFormatLoader(info.context)
-            .load((root.id, size, format))
-            .then(_resolve_background_image)
-        )
-
-    @staticmethod
-    def resolve_children(root: models.Category, info, **kwargs):
-        def slice_children_categories(children):
-            return create_connection_slice(
-                children, info, kwargs, CategoryCountableConnection
-            )
-
-        return (
-            CategoryChildrenByCategoryIdLoader(info.context)
-            .load(root.pk)
-            .then(slice_children_categories)
-        )
-
-    @staticmethod
-    def resolve_url(root: models.Category, _info):
-        return ""
-
-    @staticmethod
-    @traced_resolver
-    def resolve_products(root: models.Category, info, *, channel=None, **kwargs):
-        requestor = get_user_or_app_from_context(info.context)
-        has_required_permissions = has_one_of_permissions(
-            requestor, ALL_PRODUCTS_PERMISSIONS
-        )
-        tree = root.get_descendants(include_self=True)
-        if channel is None and not has_required_permissions:
-            channel = get_default_channel_slug_or_graphql_error()
-        qs = models.Product.objects.all()
-        if not has_required_permissions:
-            qs = (
-                qs.visible_to_user(requestor, channel)
-                .annotate_visible_in_listings(channel)
-                .exclude(
-                    visible_in_listings=False,
-                )
-            )
-        if channel and has_required_permissions:
-            qs = qs.filter(channel_listings__channel__slug=channel)
-        qs = qs.filter(category__in=tree)
-        qs = ChannelQsContext(qs=qs, channel_slug=channel)
-        return create_connection_slice(qs, info, kwargs, ProductCountableConnection)
-
-    @staticmethod
-    def __resolve_references(roots: List["Category"], _info):
-        return resolve_federation_references(Category, roots, models.Category.objects)
-
-
-class CategoryCountableConnection(CountableConnection):
-    class Meta:
-        node = Category
 
 
 @federated_entity("id")
